Fix : cart alert

+++ cart/detail.js
@@ -126,10 +126,28 @@ const cartin = (x) => {
       icon: "warning",
     });
   } else {
-    cartData.push(cartProduct);
-    localStorage.setItem("userCart", JSON.stringify(cartData));
-    updateCartCount();
-    window.location.href = "cart.html";
+    Swal.fire({
+      title:
+        '<h2 style="font-size:22px; font-weight: bold;">장바구니에 담으시겠습니까?</h2>',
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "예",
+      cancelButtonText: "아니오",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          title:
+            '<h2 style="font-size:22px; font-weight: bold;">장바구니에 담겼습니다.</h2>',
+          icon: "success",
+        });
+        cartData.push(cartProduct);
+        localStorage.setItem("userCart", JSON.stringify(cartData));
+        updateCartCount();
+        // window.location.href = "cart.html";
+      }
+    });
   }
 };
